Clarified README.md, refactored source files
Updated README.md to include the current state of the project (not
working and unfinished) and clarify project goals, as well as to
start documenting how the module works and can be used. init.lua
was getting too long so broke out some of the functions into two
submodules that contain "class" definitions for the state machine
and executable commands.