Major rewrite and README update
I ended up majorly refactoring this several times since my last
commit. Renamed a lot of things so the definition tables look less
foreign. Restructed commands to handle recursive calls (for motions)
as well as command chaining (for [[ and ]]). Added support for prompts
(as yet untested.) There is too much to list. But now, this is actually
partially working. Working features are listed in the README.